blogPost/api/serializers.py
from rest_framework import serializers
from blogPost.models import BlogPost


class BlogSerializer(serializers.ModelSerializer):
    class Meta:
        model = BlogPost
        fields = ['title', 'body', 'date_updated']