requirements.txt
Django~=3.1.6
Pillow==6.0.0
python-decouple==3.1
pytz==2019.1
sqlparse==0.3.0
djangorestframework==3.12.2